chatbot-df/requirements.txt
line-bot-sdk
flask
PyCryptodome==3.4.3
pandas
